TBR: fix build break
BUG=

Review URL: https://codereview.chromium.org//25392003

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@28075 260f80e4-7a28-3924-810f-c04153c831b5

+++ runtime/vm/object.cc
@@ -13680,7 +13680,7 @@ RawOneByteString* OneByteString::EscapeSpecialCharacters(const String& str) {
     }
     return OneByteString::raw(dststr);
   }
-  return OneByteString::New(0, Heap::kNew);
+  return OneByteString::New(static_cast<intptr_t>(0), Heap::kNew);
 }
 
 RawOneByteString* ExternalOneByteString::EscapeSpecialCharacters(
@@ -13709,7 +13709,7 @@ RawOneByteString* ExternalOneByteString::EscapeSpecialCharacters(
     }
     return OneByteString::raw(dststr);
   }
-  return OneByteString::New(0, Heap::kNew);
+  return OneByteString::New(static_cast<intptr_t>(0), Heap::kNew);
 }
